+++ src/test/java/net/jodah/failsafe/functional/PolicyOrderingTest.java
@@ -59,13 +59,14 @@ public class PolicyOrderingTest {
   }
 
   public void testExecutionWithRetryThenCircuitBreaker() {
-    RetryPolicy rp = new RetryPolicy().withMaxRetries(2);
+    RetryPolicy rp = new RetryPolicy().withMaxRetries(1);
     CircuitBreaker cb = new CircuitBreaker().withFailureThreshold(5);
 
     Execution execution = new Execution(cb, rp);
     execution.recordFailure(new Exception());
+    assertFalse(execution.isComplete());
+    execution.recordFailure(new Exception());
     assertTrue(execution.isComplete());
-
     assertTrue(cb.isClosed());
   }